chore: change duration type

+++ src/main/java/com/example/mcpclient/model/Message.java
@@ -14,7 +14,7 @@ public record Message(
         @JsonProperty("destination")
         Long destination,
         @JsonProperty("duration")
-        int duration,
+        Long duration,
         @JsonProperty("status_code")
         String statusCode,
         @JsonProperty("status_description")
